feat: Add new themes and theme provider
